Solve graduation committe problem from ncpc
Here I had to admit defeat. I am a bit worse at dynamic programming
than I thought. After a while I got a solution working, but it was too
slow. I did a really complex solution, which also did not work, using
a binary search for the correct number of questions to answer.

Finally, I looked at the solution, which was incredibly similar to my
base dp function. But, it iterated the other way than me, allowing them
to cache more things. However, this did not work when implemented,
as the hashmap was too slow, so I also had to change to arrays, which
was a bit annoying.